Separate authz code again
