+++ coderd/coderdtest/authorize_test.go
@@ -1,0 +1,20 @@
+package coderdtest_test
+
+import (
+	"context"
+	"testing"
+
+	"github.com/coder/coder/coderd/coderdtest"
+)
+
+func TestAuthorizeAllEndpoints(t *testing.T) {
+	t.Parallel()
+	client, _, api := coderdtest.NewWithAPI(t, &coderdtest.Options{
+		Authorizer:               &coderdtest.RecordingAuthorizer{},
+		IncludeProvisionerDaemon: true,
+	})
+	admin := coderdtest.CreateFirstUser(t, client)
+	a := coderdtest.NewAuthTester(context.Background(), t, client, api, admin)
+	skipRoute, assertRoute := coderdtest.AGPLRoutes(a)
+	a.Test(context.Background(), assertRoute, skipRoute)
+}
